+++ src/admin/pages/Product/ManageProductSearch.js
@@ -1,33 +1,32 @@
+import ProductShow from "../../components/Product/ProductShow"
+import ProductsContext from "../../../context/product";
 import React, { useState, useContext, useEffect } from 'react';
-import ProductShow from "../components/Product/ProductShow"
-import ProductsContext from "../context/product";
-import Avatar from "../../src/assets/images/avatar-photo.jpg";
-import { useNavigate } from 'react-router-dom';
-import useKeyPress from "../hooks/use-key-press";
-import Layout from './components/Layout';
+import Avatar from "../../../assets/images/avatar-photo.jpg";
+import useKeyPress from "../../../hooks/use-key-press";
+import { useSearchParams, useNavigate } from 'react-router-dom';
+import Layout from "../../components/Layout";
 
-function ManageProduct() {
+function ManageProductSearch() {
     const navigate = useNavigate("");
-    const [searchString, setSearchString] = useState("");
-
+    const { searchProduct } = useContext(ProductsContext);
+    const [searchParams] = useSearchParams();
+    const searchQuery = searchParams.get('keyword') || '';
+    const [searchString, setSearchString] = useState(searchQuery);
+    useEffect(() => {
+        searchProduct(searchQuery);
+    }, [searchQuery]);
     const handleSearch = async () => {
-        if (searchString !== '') {
+        if(searchString !== ''){
             const encodedQuery = encodeURIComponent(searchString);
-            console.log(encodedQuery)
             navigate(`/admin/manage-product/search?keyword=${encodedQuery}`);
+        } else {
+            navigate("admin/manage-product");
         }
     };
-
-    const { fetchProducts } = useContext(ProductsContext);
-    useEffect(() => {
-        fetchProducts();
-    }, []);
-
     useKeyPress(handleSearch)
-
     return (
-        <Layout title="Quản lý sản phẩm" activePage="product">
-            <div className="flex justify-start">    
+        <Layout activePage="product">
+            <div className="flex justify-start">
                 <div className="w-full">
                     <div className="bg-stone-100 py-2">
                         <header className="mx-auto flex h-16 max-w-[1200px] items-center justify-between px-5">
@@ -46,7 +45,6 @@ function ManageProduct() {
                                         d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607z"
                                     />
                                 </svg>
-
                                 <input
                                     className="hidden w-11/12 outline-none md:block"
                                     type="search"
@@ -54,7 +52,6 @@ function ManageProduct() {
                                     value={searchString}
                                     onChange={(e) => setSearchString(e.target.value)}
                                 />
-
                                 <button
                                     className="ml-auto h-full bg-amber-400 px-4 hover:bg-yellow-300 rounded-r-lg text-white"
                                     onClick={handleSearch}
@@ -62,15 +59,14 @@ function ManageProduct() {
                                     Tìm
                                 </button>
                             </div>
-
-                            <div className="flex items-center">
-                                <div className="relative">
-                                    <a className="flex items-center" href="#">
-                                        <span className="hidden text-right lg:block">
-                                            <span className="block text-sm font-medium text-black dark:text-white">Quản trị viên</span>
+                            <div class="flex items-center">
+                                <div class="relative">
+                                    <a class="flex items-center" href="#">
+                                        <span class="hidden text-right lg:block">
+                                            <span class="block text-sm font-medium text-black dark:text-white">Quản trị viên</span>
                                         </span>
 
-                                        <span className="h-12 w-12 rounded-full ml-2">
+                                        <span class="h-12 w-12 rounded-full ml-2">
                                             <img src={Avatar} alt="User" />
                                         </span>
                                     </a>
@@ -88,4 +84,4 @@ function ManageProduct() {
     )
 }
 
-export default ManageProduct;
+export default ManageProductSearch;
